requirements.txt
Adafruit-Blinka==8.66.2
Adafruit-Blinka-Raspberry-Pi5-Neopixel==1.0.0rc2
adafruit-circuitpython-busdevice==5.2.13
adafruit-circuitpython-connectionmanager==3.1.5
adafruit-circuitpython-dht==4.0.9
adafruit-circuitpython-requests==4.1.13
adafruit-circuitpython-typing==1.12.2
Adafruit-PlatformDetect==3.84.0
Adafruit-PureIO==1.1.11
async-timeout==5.0.1
autocommand==2.2.2
backports.tarfile==1.2.0
binho-host-adapter==0.1.6
blinker==1.5
certifi==2025.10.5
cffi==2.0.0
chardet==5.1.0
charset-normalizer==3.4.3
cheroot==11.0.0
CherryPy==18.10.0
coloredlogs==15.0.1
colorzero==2.0
cryptography==38.0.4
dbus-python==1.3.2
distro==1.8.0
filelock==3.19.1
flatbuffers==20181003210633
fsspec==2025.9.0
gpiozero==2.0.1
h5py==3.14.0
hf-xet==1.1.10
httplib2==0.20.4
huggingface_hub==0.35.3
humanfriendly==10.0
idna==3.10
jaraco.collections==5.2.1
jaraco.context==6.0.1
jaraco.functools==4.3.0
jaraco.text==4.0.0
Jinja2==3.1.6
lazr.restfulclient==0.14.5
lazr.uri==1.0.6
lgpio==0.2.2.0
MarkupSafe==3.0.3
more-itertools==10.8.0
mpmath==1.3.0
networkx==3.5
numpy==2.3.3
oauthlib==3.2.2
onnxruntime==1.22.0
packaging==25.0
paho-mqtt==2.1.0
pandas==2.3.3
pigpio==1.78
portend==3.2.1
protobuf==6.32.1
pycparser==2.23
pycryptodomex==3.11.0
pyftdi==0.57.1
PyGObject==3.42.2
PyJWT==2.6.0
pyparsing==3.0.9
pyserial==3.5
python-apt==2.6.0
python-dateutil==2.9.0.post0
pytz==2025.2
pyusb==1.3.1
PyYAML==6.0.3
redis==6.4.0
regex==2025.9.18
requests==2.32.5
rpi-keyboard-config==1.0
rpi-lgpio==0.6
rpi-ws281x==5.0.0
RPi.GPIO==0.7.1
safetensors==0.6.2
scipy==1.16.2
six==1.17.0
smbus2==0.4.2
sounddevice==0.5.2
spidev==3.5
ssh-import-id==5.10
sympy==1.14.0
sysv-ipc==1.1.0
tempora==5.8.1
tokenizers==0.22.1
toml==0.10.2
torch==2.7.0
torchaudio==2.7.0
tqdm==4.67.1
transformers==4.57.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wadllib==1.3.6
zc.lockfile==4.0
